services/binance-service/pyproject.toml
[project]
name = "binance-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncpg>=0.31.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "websockets>=16.0",
    "cryptography>=44.0.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
